created folder for user profile page
